+++ src/app/features/events/events-fav/events-fav.component.ts
@@ -1,8 +1,5 @@
-import { Component, ViewChild, OnInit, OnDestroy } from '@angular/core';
+import { Component, OnInit, OnDestroy } from '@angular/core';
 import { Router } from '@angular/router';
-import { MatPaginator } from '@angular/material/paginator';
-import { MatSort } from '@angular/material/sort';
-import { MatTableDataSource } from '@angular/material/table';
 
 import { AngularFireAuth } from '@angular/fire/compat/auth';
 import { Observable, Subscription } from 'rxjs';
@@ -10,12 +7,10 @@ import { map } from 'rxjs/operators';
 import { formatDistance } from 'date-fns';
 import { es } from 'date-fns/locale';
 
-import { AuthService } from '@auth/auth.service';
 import { IEntity } from '@models/entity';
 import { IEvent } from '@models/event';
 import { IUser } from '@models/user';
 import { UserService } from '@services/users.service';
-import { LogService } from '@services/log.service';
 import { EventService } from '@services/events.service';
 import { EntityService } from '@services/entities.service';
 import { SpinnerService } from '@services/spinner.service';
@@ -27,24 +22,18 @@ import { SpinnerService } from '@services/spinner.service';
 })
 export class EventsFavComponent implements OnInit, OnDestroy {
 
-  @ViewChild(MatPaginator, {static: false}) paginator: MatPaginator;
-  @ViewChild(MatSort, {static: false}) sort: MatSort;
-
   public viewMode = 'cards';
   public events: IEvent[] = [];
   public favEntitiesStr: string[] = [];
   public favEntities$: Observable<IEntity[]>;
-  public dataSource: MatTableDataSource<IEvent> = new MatTableDataSource();
   public displayedColumns: string[] = [ 'status', 'id', 'timestamp', 'image', 'collapsed-info', 'name', 'categories', 'actions1'];
-  private userLogged: IUser;
+  public userLogged: IUser;
   private listOfObservers: Array<Subscription> = [];
 
   constructor(
     private router: Router,
     public auth: AngularFireAuth,
-    private authSrv: AuthService,
     private userSrv: UserService,
-    private logSrv: LogService,
     private spinnerSvc: SpinnerService,
     private eventSrv: EventService,
     private entitySrv: EntityService
@@ -76,16 +65,10 @@ export class EventsFavComponent implements OnInit, OnDestroy {
             )
             .subscribe((events: IEvent[]) => {
               this.events = events;
-              this.dataSource = new MatTableDataSource(this.events);
               this.spinnerSvc.hide();
-              this.dataSource.paginator = this.paginator;
-              this.dataSource.sort = this.sort;
             });
         } else {
-          this.dataSource = new MatTableDataSource(this.events);
           this.spinnerSvc.hide();
-          this.dataSource.paginator = this.paginator;
-          this.dataSource.sort = this.sort;
         }
 
         // Entities
@@ -99,29 +82,12 @@ export class EventsFavComponent implements OnInit, OnDestroy {
     this.listOfObservers.push(subs1$);
   }
 
-  applyFilter(filterValue: string): void {
-    this.dataSource.filter = filterValue.trim().toLowerCase();
-
-    if (this.dataSource.paginator) {
-      this.dataSource.paginator.firstPage();
-    }
-  }
-
-  public gotoItem(event: IEvent): void {
-    this.router.navigate([`eventos/${event.id}`]);
-  }
-
   public gotoEntity(entity: IEntity): void {
     this.router.navigate([`entidades/${entity.id}`]);
   }
 
   public gotoEntitiesFav(): void {
     this.router.navigate([`entidades/favoritas`]);
-  }
-
-  public removeFav(event: IEvent): void {
-    this.userLogged.favEvents = this.userLogged.favEvents.filter( (eventId: string) => eventId !== event.id );
-    this.userSrv.updateUser(this.userLogged);
   }
 
   ngOnDestroy(): void {